fixed Beksultans bugs

+++ trading/utils.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
 
-
-def get_commission(sum: int, percent: str) -> Decimal:
-    return sum - (sum/100 * Decimal(percent))
+def get_commission(sum: int, percent: str) -> float:
+    return sum - (sum/100 * float(percent))
